+++ src/reg_intf_pkg.sv
@@ -9,6 +9,7 @@
 // specific language governing permissions and limitations under the License.
 //
 // Florian Zaruba <zarubaf@iis.ee.ethz.ch>
+// Fabian Schuiki <fschuiki@iis.ee.ethz.ch>
 
 /// A simple register interface.
 ///
@@ -19,7 +20,7 @@
 /// and ready are high. Valid must not depend on ready. The slave presents the
 /// read data and error signals. These signals must be constant while valid and
 /// ready are both high.
-package reg_intf;
+package reg_intf_pkg;
 
     /// 32 bit address, 32 bit data request package
     typedef struct packed {
@@ -28,7 +29,7 @@ package reg_intf;
         logic [31:0] wdata;
         logic [3:0]  wstrb;
         logic        valid;
-    } reg_intf_req_a32_d32;
+    } req_a32_d32;
 
     /// 32 bit address, 64 bit data request package
     typedef struct packed {
@@ -37,20 +38,20 @@ package reg_intf;
         logic [63:0] wdata;
         logic [7:0]  wstrb;
         logic        valid;
-    } reg_intf_req_a32_d64;
+    } req_a32_d64;
 
     /// 32 bit Response packages
     typedef struct packed {
         logic [31:0] rdata;
         logic        error;
         logic        ready;
-    } reg_intf_resp_d32;
+    } rsp_d32;
 
     /// 32 bit Response packages
     typedef struct packed {
         logic [63:0] rdata;
         logic        error;
         logic        ready;
-    } reg_intf_resp_d64;
+    } rsp_d64;
 
 endpackage